+++ route/darwin.go
@@ -1,0 +1,22 @@
+//go:build darwin
+// +build darwin
+
+/*
+ * @Author: LinkLeong link@icewhale.org
+ * @Date: 2022-08-12 14:22:28
+ * @LastEditors: LinkLeong
+ * @LastEditTime: 2022-08-12 14:22:44
+ * @FilePath: /CasaOS/route/darwin.go
+ * @Description:
+ * @Website: https://www.casaos.io
+ * Copyright (c) 2022 by icewhale, All Rights Reserved.
+ */
+
+package route
+
+func MonitoryUSB() {
+
+}
+func SendAllHardwareStatusBySocket() {
+
+}