+++ route/periodical.go
@@ -1,8 +1,11 @@
+//go:build !darwin
+// +build !darwin
+
 /*
  * @Author: LinkLeong link@icewhale.com
  * @Date: 2022-07-01 15:11:36
  * @LastEditors: LinkLeong
- * @LastEditTime: 2022-08-10 17:19:27
+ * @LastEditTime: 2022-08-12 14:19:53
  * @FilePath: /CasaOS/route/periodical.go
  * @Description:
  * @Website: https://www.casaos.io