+++ route/darwin.go
@@ -5,7 +5,7 @@
  * @Author: LinkLeong link@icewhale.org
  * @Date: 2022-08-12 14:22:28
  * @LastEditors: LinkLeong
- * @LastEditTime: 2022-08-12 14:22:44
+ * @LastEditTime: 2022-08-12 18:41:14
  * @FilePath: /CasaOS/route/darwin.go
  * @Description:
  * @Website: https://www.casaos.io
@@ -18,5 +18,8 @@ func MonitoryUSB() {
 
 }
 func SendAllHardwareStatusBySocket() {
+
+}
+func SendUSBBySocket() {
 
 }