+++ route/v1/system.go
@@ -182,7 +182,31 @@ func PutSystemUSBAutoMount(c *gin.Context) {
 		service.MyService.System().UpdateUSBAutoMount("False")
 		service.MyService.System().ExecUSBAutoMountShell("False")
 	}
+	go func() {
+		usbList := service.MyService.Disk().LSBLK(false)
+		usb := []model.DriveUSB{}
+		for _, v := range usbList {
+			if v.Tran == "usb" {
+				isMount := false
+				temp := model.DriveUSB{}
+				temp.Model = v.Model
+				temp.Name = v.Name
+				temp.Size = v.Size
+				for _, child := range v.Children {
+					if len(child.MountPoint) > 0 {
+						isMount = true
+						avail, _ := strconv.ParseUint(child.FSAvail, 10, 64)
+						temp.Avail += avail
 
+					}
+				}
+				if isMount {
+					usb = append(usb, temp)
+				}
+			}
+		}
+		service.MyService.Notify().SendUSBInfoBySocket(usb)
+	}()
 	c.JSON(common_err.SUCCESS,
 		model.Result{
 			Success: common_err.SUCCESS,
@@ -202,7 +226,31 @@ func GetSystemUSBAutoMount(c *gin.Context) {
 	if config.ServerInfo.USBAutoMount == "False" {
 		state = "False"
 	}
+	go func() {
+		usbList := service.MyService.Disk().LSBLK(false)
+		usb := []model.DriveUSB{}
+		for _, v := range usbList {
+			if v.Tran == "usb" {
+				isMount := false
+				temp := model.DriveUSB{}
+				temp.Model = v.Model
+				temp.Name = v.Name
+				temp.Size = v.Size
+				for _, child := range v.Children {
+					if len(child.MountPoint) > 0 {
+						isMount = true
+						avail, _ := strconv.ParseUint(child.FSAvail, 10, 64)
+						temp.Avail += avail
 
+					}
+				}
+				if isMount {
+					usb = append(usb, temp)
+				}
+			}
+		}
+		service.MyService.Notify().SendUSBInfoBySocket(usb)
+	}()
 	c.JSON(common_err.SUCCESS,
 		model.Result{
 			Success: common_err.SUCCESS,